When a request completion is received in lock_dlm during recovery, the lock must be demoted to NL and rerequested in the originally requested mode after recovery completes (only NOEXP requests can be completed for gfs during recovery.)
When a lock was rerequested, the QUECVT flag was always being set even
though the lock may also have the EXPEDITE (gfs's PRIORITY) flag set.
It's illegal to use EXPEDITE and QUECVT together, so we now only set
QUECVT if EXPEDITE isn't.  An assertion fails if both flags are used.