By default, gulm determins the name of the machine it is on with gethostbyname(). Then it looks up an ip from this name to use as its IP.
There are now three ways to override these defaults.  --name will use the
name given instead of what is in gethostbyname().  --ip will use that
address for this machine. (don't be stupid and set it to a value that isn't
configured to this machine.)  And --ifdev will lookup the ip from the
configured network device.  --ip overrides --ifdev.

These options are only available form the command line options.