Partial workthrough of log messages.  Some moving, some changing. mostly changed `errors' that gulm can handle to not use log_err
Some corrections to the estimated lock space size calculation.
Still needs more work, but its closer.  Once this gets more accurate
will use it for th highwater mark instead of lock count. But not yet.

Added out going queues to the local connect in LTPX. Fixes bug 146670.
Might need to add read penalties, but works for me without, so holding
off on that.

Removed a util_* that hasn't been used in quite some time, but was
still being compiled and linked in.