- fence_node has been change to work better with gulm (143487) - Two new options have been added   -c <cluster_name>   -O
The -O overrides the quorum requirement.  It requires the use of -c,
due to the fact that CCS may be in a transitional phase that only
solidifies when the cluster is quorate.

Although I haven't tested it, this may allow a person to
fence_node -O -c someone_else_cluster someone_else_node... a
feature? a bug?