Improve logic that delays and reduces fencing.  When fenced is recovering for a failed node, the 'post_fail_delay' is used to give victims some time to rejoin the cluster and avoid being fenced.  If this happens once, then it's likely to happen again and the 'post_join_delay' is more appropriate, so fenced switches to the 'post_join_delay' value (if it's larger which is usually the case.)
The common situation where this helps is when multiple nodes fail causing
the cluster to lose quorum and then the failed nodes all rejoin the
cluster at about the same time.  The rejoining nodes are more likely
to all avoid being fenced if fenced uses the larger post_join_delay.