- fix for bug 157094
A mysterious error being generated when trying to do a broadcast (sendto):
ccsd[1704]: Unable to perform sendto: Cannot assign requested address

On certain clusters (seems to be when ccs tries using IPv6), this error
could show up 9 out of 10 times.  When the error was received, the
broadcast attempt would fail.  This caused the attempt to grab any
possibly updated cluster.conf files to abort.

Waiting a moment, closing the socket, reopening the socket, and retrying
the broadcast seems to solve the issue.  (It has work 100+ times so far.)

I'm not entirely certain what is causing the initial try to fail - perhaps
the underlying subsystem is not quite ready...  In any case, I have never
seen a second attempt fail.