fix for bz 161352
Adds support for latest ilo firmware version (1.75).  Changes were also
added to make sure that power status of the machine is being properlly
checked after power change commands have been issued.  Before, it was
assumed to work through the messages recieved in the xml response.  As a
result, systems with APCI  enabled would soft power off leaving, systems
running instead of properlly fencing them.

A new option, ribcl,  was added to the agent to force the protocol to use
when trying to send the power off command.  Setting this value to "2.21"
might be required to work around some versions of ilo on machines with APCI
enabled.  If left blank, the protocol version is autodetected.