N.B. This patch changes the ondisk format of GFS2, so you'll need to remake filesystems.
The idea here is to make the metadata closer to that of GFS1 in order to
reduce the amount of work required to migrate GFS1 -> GFS2. This patch
changes the numercial value of a few constants to match those of
identical function in GFS1. One of the GFS1 constants is no longer
used in GFS2 (as noted in the comments) and there are two new constants
added.

As far as I can tell there are no more constants in GFS2 which are
not comaptible with those in GFS1. All the remaining ondisk
incompatibilities are due to structures changing size and I'll
be taking a look at those next.