Sorry - another ondisk format changing check in...
I've reordered the fields in the common metadata header and added a pad
field so that this structure is now of identical size as in gfs1. It
also has identically functioning fields in the same place.

One new field is added, mh_blkno which contains the block number of
the block in question. This is instead of the generation number used
in gfs1's journaling and now no longer needed. The mh_blkno field is
checked only once in the code and could be moved into the structure
where its required, rather than being left in the common header. (I
checked this with Ken who added the field in the first place)

I have a thought that it might be a useful feature to have all ondisk
metadata checksummed in which case this header field might be just the
right place to put it.

Also some padding is added to the dinode, sb and rgrp structures in order
to make them compatible with the earier gfs1 versions. Again fields are
now in the same places as in gfs1 with padding added where required since
the fields have got smaller.

I don't believe that there will be any significant performance impact in
changing these structures, but it should make things much easier when it
comes to migration.

There are other structures (I think that the various directory metadata
comes under this heading) where removing the padding is likely to
improve performance significantly and so I'm intending to take a
different approach there.

There is likely to be at least one more metadata change for gfs2 in
the not too distant future. My next main concern is to make gfs2
bigendian on disk, to match gfs1. Beyond that, all the other changes
should be much easier to cope with. I did a test recently and found that
given a few 32bit fields to copy, adding endian coversion made about
12% difference to the speeed of copying (which we are doing anyway).

If we keep the two versions with the same on disk endianess then a lot
of the metadata will not need conversion at all (given the changes in
this patch). So an upgrade will be quite quick, rather than needing to
rewrite each and every metadata disk block.