Write a list of the block numbers of the succeding metadata blocks which are about to be committed into the log. This uses the spare space at the end of the log descriptor block, so mostly it doesn't actually require any more disk I/O to do this, and when it does, its very little.
This doesn't actually change recovery (yet) since the recovery side is
still making use of the mh_blkno field in the metadata header still. The
intention is to use the list of block numbers instead. This has a slight
speed advantage in recovery since we no longer need to read any block
which is in the revoke list, since we will have already read a list of
the block numbers ahead of time.

Once this is done, we'll no longer need to use the mh_blkno field in
the common metadata header which gives two advantages:

 o We don't need the bmap ioctl() since its only use is the online version
of mkfs
 o We don't need to rewrite all the metadata headers when converting gfs1
to gfs2

I plan to use an almost identical system to deal with data blocks in
the journal, with the addition of an extra field to flag whether the
data block begins with the GFS magic number or not, which is required
so that we can escape any data blocks which do start with the magic
number. When thats done, it will greatly simplify dealing with
journaled data files.