This takes mkfs out of the main build system. It should make it much easier to build it since its just a question of editing the Makefile to point at the kernel source and then, make, make install.
Also the binary will be called mkfs.gfs2 and make install will stick
it in /sbin so that you can call it through the mkfs generic front
end like all the other filesystems do.

So hopefully this will be easier to build and easier to use. It also
incorporates the changes required to make filesystems for the new
journaled file format, so be aware that only the very latest git
tree of gfs2 will work with it.