Added success and failure commands in start/stop.
scsi_reserve start - success if we can register with a device.
scsi_reserve stop  - success if wa can unregister with a device.

Note that we always try to create a reservation for a device in start,
but we do not care about success/failure. If it fails, it is probably
because a reservation already exists.