Add plock rate limit option -l <limit>.  Current default is no limit (0). If a limit is set, gfs_controld will send no more than <limit> plock operations (multicast messages) every second.
Given a limit of 10, one file system where plocks are used, and a program
that does a tight loop of fcntl lock/unlock operations, the max number of
loop iterations in 1 second would be 5.  If eight nodes were all doing
this there would be 80 total network multicasts every second from all
nodes in the cluster.

We also record the volume of plock messages accepted locally and received
from the network in the debug log.  A log entry is written for every
1000 locally accepted plock operations and for every 1000 operations
received from the network.