Bug 236580: [HA LVM]: Bringing site back on-line after failure causes pr...
Setup:
- 2 interconnected sites
- each site has a disk and a machine
- LVM mirroring is used to mirror the disks from the sites

When one site fails, the LVM happily moves over to the second site -
removing the failed disk from the VG that was part of the failed
site.  However, when the failed site is restored and the service
attempts to move back to the original machine, it fails because
of the conflicts in LVM metadata on the disks.

This fix allows the LV to be reactivated on the original node
by filtering out the devices which have stale metadata (i.e
the device that was removed during the failure).