RedHat bugzilla 239727:
Previous CVS check-in did a last minute change with the way purge count
was calculated. The intention was to trim glocks evenly across all the
hash buckets and apparently the size of hash array was overlooked. It
ends up with zero trimming count most of the time. This virtually makes
glock trimming patch a void feature.