Added ability to parse and print journal information.  For example: gfs2_edit -p journal2 /dev/roth_vg/roth_lv
Also added the ability to jump relative block numbers.  This is for
when you jump to a block by "editing" the block number:
Cursor up to block number, press <enter> to enter destination block.
Before you could type a structure or block number.  Now you can type
in a structure, block number, or relative block number.  For example,
if you're on the superblock (block 0x10) enter +0x300 will take you
to block 0x310.  This number may be in decimal or hex and may be
positive or negative.