Detection and resolution now works with my basic deadlock tests.
Had to rework how lock state is assembled.  The previous method was
simpler and just gathered the master lock state from all nodes, but
I failed to realize that the xid (transaction id) isn't synced to
remote master copy locks.  So, now all lock state is saved in the
checkpoints, both master-copy and process-copy (containing the xid)
which are then merged to give the full view of the lock.