[CMAN] Don't ignore cman_tool version
the processing of the 'cman_tool version' command updates the
config_version variable immediately rather than waiting for
the message to come back. Because of this, when the message
/does/ come back it gets ignored because cman thinks it
already has the latest config!

The solution is, of course, not to update the config version
until we receive the RECONFIGURE message.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>