[CMAN] valid port number & don't use it before validation
cman_send_data didn't validate the port number. This was
not a great problem, but it was slightly silly.

cman_start_recv_data used the port number before validation
in a debug message, this could crash the server if debug
is enabled and a very large port number was passed in.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>