rgmanager/lvm.sh:  Fix bug bz242798
Allow a machine to fence itself in the event that it
cannot deactivate logical volumes.  (The user must
explicitly enable this option.)  This is useful in cases
where one machine in the cluster looses connectivity to
its resources, but the others don't.  The machine fences
itself and the service moves to another machine.