[CMAN] Save the new expected_votes when a node is removed
When a node leaves the cluster using 'cman_tool leave remove' it reduces
the quorum of the cluster to accomodate the loss of the node. Unfortunately
the following transition messages raise quorum back to its original value
again because of the bug fix for bz#308581, so it appears that the
remove hasn't worked.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>