[CCS] Detach dependency on ccsd to run the cluster
The old static libccs is now called libccscompat and should not be used
by anything outside ccs/. Note that the headers and the library are not
installed, nor configure knows about it. All application paths that require
compat are made static and non-configurable.

Allow ccsais and ccs_tool to use libccscompat.

Stop linking ccsd with libccs. It was not required before either.

Add new shared library called libccs that uses aisexec confdb directly.
The new libccs retains API compatibility with the old library so it was
possible to "unplug" the old and "plug" the new one in one go. All the
applications in the stack are now quering aisexec db directly.

NOTES:
 - the new library does not emulate all the libccscompat calls as some
   of them are not used anywhere but this is not a problem since libccs
   was static. We are shipping it now shared and we a new soname.
 - the library implements only a small subset of xpath to query the aisexec
   config db.

Port ccs_test to use the new libccs and drop all calls that are not used anywhere.

Set new default ccsincdir and ccslibdir to ccs/libccsconfdb in configure.

NOTE to developers: you will need to rerun configure.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>