gnbd-kernel: Fix receiver race
It is possible to have the gnbd receiver process finish and end a request before
the sending process has finished using the request structure.  This can cause a
kernel panic.

This fix adds a waitqueue (tx_wait) and a pointer to the request currently
being send (current_request) to the gnbd device structure.  current_request is
set before any request is sent to the server. When the send is complete, it is
cleared and the wait_queue is woken.  A new function, wait_for_send() is called
whenever it is possible for a call to gnbd_end_request() to interleave with a
send. It waits on the waitqueue if the request about to be ended is currently
being sent.

Conflicts:

	gnbd-kernel/src/gnbd.c