[CONFIG] Add ldap configurator
This is an openais configuration plugin to read the cluster config
from an LDAP server. A schema file is included that provides just
enough information to get a cluster running, more will follow.

There is also an example ldif file to show how to load the information
into the database.

The defaults are slightly odd at the moment, I'll fix those as it develops
and document how to override them. In the mean time see the top of the
source code.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>