[FENCE] Bug #448822: fence_ilo doesn't work with iLO
New fencing agent for iLO used ssh/telnet to connect, but unfortutely
there is a problem with power off. This is why we need to use SSL connection
and RIBCL commands. As there is no (?) telnet with ssl connection in RHEL we
need one to be able to use same infrastructure as in other agents. This agent
was not tested with RIBCL version < 2.0 (these part where just ported from the
old perl fencing agent)

@todo: we have to put telnet_ssl.py somewhere, I'm not sure where