[CONFIG] Add ldap loader
Add a new tool to load an existing cluster config into LDAP. This
is as incomplete as the schema at the moment, but it will allow
you to add nodes and some very basic fencing devices.

It uses libconfdb so it can migrate either a running system
or a cluster.conf file.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>