[CCS] Fold ccs_test into ccs_tool and tidy
Move all the code from ccs_test into ccs_tool and make it
switch depending on argv[0]. Also add ccs_tool query that
does full XPath queries on confdb.

When build with LEGACY_CODE set ccs_tool will behave as it
used to, talking to ccsd. Without that it will use the
new confdb system and assume that ccsd is NOT used for
distribution of cluster configurations.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>