fence: port scsi agent to use ccs_tool query and drop XML::LibXML requirement
The new ccs_tool query is able to answer to the queries from
fence_scsi*.

Drop the need of XML::LibXML.

Use ccs_tool query directly.

Abstract from the need of parsing cluster.conf that is not the
authoritative config file anymore.

Drop unrequire check_config_nodes that's already embedded in the need of
cman to run this code.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>